runpod/requirements.txt
pillow
xformers==0.0.24
torch==2.2.0
torchvision
transformers
huggingface-hub
chainlit
pydantic
timm
loguru
einops
sse-starlette